+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## [0.0.1] - TODO: Add release date.
+## [3.0.1] - TODO: Add release date.
 
-* TODO: Describe initial release.
+* 添加了初始页面及参数 
+
+  final String appName;     // 应用名
+  final String description; // 应用描述
+  final Color background;   // 欢迎页面背景色
+  final IconData icon;      // 图标
+  
+